# internal/server/testdata/openapi3.json
{
  "openapi": "3.0.0",
  "components": {
    "schemas": {
...
      "EmptyResponse": {},
      "Error": {
        "properties": {
          "code": {
            "format": "int32",
            "type": "integer"
          },
          "fieldErrors": {
            "items": {
              "properties": {
                "errors": {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                "fieldName": {
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "ListResponse_AccessKey": {
        "properties": {
          "count": {
            "format": "int",
            "type": "integer"
          },
          "items": {
            "items": {
              "properties": {
                "created": {
                  "description": "formatted as an RFC3339 date-time",
                  "example": "2022-03-14T09:48:00Z",
                  "format": "date-time",
                  "type": "string"
                },
                "expires": {
                  "description": "key is no longer valid after this time",
                  "example": "2022-03-14T09:48:00Z",
                  "format": "date-time",
                  "type": "string"
                },
                "extensionDeadline": {
                  "description": "key must be used within this duration to remain valid",
                  "example": "2022-03-14T09:48:00Z",
                  "format": "date-time",
                  "type": "string"
                },
                "id": {
                  "example": "4yJ3n3D8E2",
                  "format": "uid",
                  "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
                  "type": "string"
                },
                "issuedFor": {
                  "example": "4yJ3n3D8E2",
                  "format": "uid",
                  "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
                  "type": "string"
                },
                "issuedForName": {
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "providerID": {
                  "example": "4yJ3n3D8E2",
                  "format": "uid",
                  "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
                  "type": "string"
                }
              },
              "type": "object"
            },
            "type": "array"
          },
          "limit": {
            "format": "int",
            "type": "integer"
          },
          "page": {
            "format": "int",
            "type": "integer"
          },
          "totalCount": {
            "format": "int",
            "type": "integer"
          },
          "totalPages": {
            "format": "int",
            "type": "integer"
          }
        }
      },
...
    }
  },
  "info": {
    "description": "Infra API",
    "license": {
      "name": "Elastic License v2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    },
    "title": "Infra API",
    "version": "0.0.0"
  },
  "paths": {
    "/api/access-keys": {
      "get": {
        "description": "ListAccessKeys",
        "operationId": "ListAccessKeys",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "schema": {
              "example": "4yJ3n3D8E2",
              "format": "uid",
              "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "show_expired",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "format": "int",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int",
              "maximum": 1000,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_AccessKey"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "ListAccessKeys",
        "tags": [
          "Authentication"
        ]
      },
      "post": {
        "description": "CreateAccessKey",
        "operationId": "CreateAccessKey",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "extensionDeadline": {
                    "description": "How long the key is active for before it needs to be renewed. The access key must be used within this amount of time to renew validity",
                    "example": "72h3m6.5s",
                    "format": "duration",
                    "type": "string"
                  },
                  "name": {
                    "format": "[a-zA-Z0-9\\-_.]",
                    "maxLength": 256,
                    "minLength": 2,
                    "type": "string"
                  },
                  "ttl": {
                    "description": "maximum time valid",
                    "example": "72h3m6.5s",
                    "format": "duration",
                    "type": "string"
                  },
                  "userID": {
                    "example": "4yJ3n3D8E2",
                    "format": "uid",
                    "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
                    "type": "string"
                  }
                },
                "required": [
                  "userID",
                  "ttl",
                  "extensionDeadline"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateAccessKeyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "CreateAccessKey",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/access-keys/{id}": {
      "delete": {
        "description": "DeleteAccessKey",
        "operationId": "DeleteAccessKey",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "4yJ3n3D8E2",
              "format": "uid",
              "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "DeleteAccessKey",
        "tags": [
          "Authentication"
        ]
      }
    },
    "/api/destinations": {
      "get": {
        "description": "ListDestinations",
        "operationId": "ListDestinations",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "name",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "unique_id",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "format": "int",
              "minimum": 0,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "format": "int",
              "maximum": 1000,
              "minimum": 0,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListResponse_Destination"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "ListDestinations",
        "tags": [
          "Destinations"
        ]
      },
      "post": {
        "description": "CreateDestination",
        "operationId": "CreateDestination",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "connection": {
                    "properties": {
                      "ca": {
                        "example": "-----BEGIN CERTIFICATE-----\nMIIDNTCCAh2gAwIBAgIRALRetnpcTo9O3V2fAK3ix+c\n-----END CERTIFICATE-----\n",
                        "type": "string"
                      },
                      "url": {
                        "example": "aa60eexample.us-west-2.elb.amazonaws.com",
                        "type": "string"
                      }
                    },
                    "required": [
                      "url"
                    ],
                    "type": "object"
                  },
                  "name": {
                    "format": "[a-zA-Z0-9\\-_.]",
                    "maxLength": 256,
                    "minLength": 2,
                    "type": "string"
                  },
                  "resources": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "roles": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  },
                  "uniqueID": {
                    "type": "string"
                  },
                  "version": {
                    "type": "string"
                  }
                },
                "required": [
                  "uniqueID",
                  "name"
                ],
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Destination"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "CreateDestination",
        "tags": [
          "Destinations"
        ]
      }
    },
...
    "/api/grants/{id}": {
      "delete": {
        "description": "DeleteGrant",
        "operationId": "DeleteGrant",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "4yJ3n3D8E2",
              "format": "uid",
              "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmptyResponse"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "DeleteGrant",
        "tags": [
          "Grants"
        ]
      },
      "get": {
        "description": "GetGrant",
        "operationId": "GetGrant",
        "parameters": [
          {
            "in": "header",
            "name": "Infra-Version",
            "required": true,
            "schema": {
              "description": "Version of the API being requested",
              "example": "0.0.0",
              "format": "\\d+\\.\\d+\\(.\\d+)?(-.\\w(+\\w)?)?",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "example": "4yJ3n3D8E2",
              "format": "uid",
              "pattern": "[\\da-zA-HJ-NP-Z]{1,11}",
              "type": "string"
            }
          }
        ],
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Unauthorized: Requestor is not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Forbidden: Requestor does not have the right permissions"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Duplicate Record"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Grant"
                }
              }
            },
            "description": "Success"
          }
        },
        "summary": "GetGrant",
        "tags": [
          "Grants"
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://api.infrahq.com"
    }
  ]
}